pyproject.toml
[project]
name = "week-6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "peft>=0.15.1",
    "pillow>=11.1.0",
    "python-multipart>=0.0.20",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.1",
    "uvicorn>=0.34.0",
]

[tool.ruff]
indent-width = 2
line-length = 120
select = ["E", "W", "F"]

[tool.ruff.lint]
extend-select = ["E", "W", "F"]
